requirements.txt
astroid
async-timeout
attrs
autobahn
Automat
autopep8
beautifulsoup4
bs4
certifi
cffi
channels
channels-redis
charset-normalizer
constantly
cryptography
daphne
Deprecated
dill
Django~=5.0.6
django-cors-headers
django-environ~=0.11.2
django-filter~=24.2
djangorestframework~=3.15.1
djangorestframework-jwt~=1.11.0
flake8
hyperlink
idna
incremental
isort
lazy-object-proxy
mccabe
msgpack
packaging
Pillow
platformdirs
psycopg2-binary
pyasn1
pyasn1-modules
pycodestyle
pycparser
pyflakes
PyJWT
pylint
pyOpenSSL
pyparsing
python-dateutil
pytz
redis
requests~=2.32.3
service-identity
six
soupsieve
sqlparse
stripe
tomli
tomlkit
Twisted
txaio
typing_extensions
urllib3
wget

drf-yasg~=1.21.7
djangorestframework-simplejwt~=5.3.1